# api/api.yaml
swagger: '2.0'
...
info:
  version: "1.0.0"
  title: ASW Issue Tracker API
  description: |
    The first version of the ASW API is an exciting step forward towards
    making it easier for users to have open access to their data. We created it
    so that you can surface the amazing content ASW-Project users share.

    Build something great!

    All endpoints are only accessible via http and are located at
    `api.aswproject.fun`. 
    
    You must use an API Key for each endpoint, the new user creation *POST /users* doesn't require it though.
    
    In order to get yours, you should LogIn here [api.aswproject.fun/login](http://api.aswproject.fun/login), once you get authenticated with your email and password, a message will appear telling you your API Key, copy it.
    To use your API Key, click on the green Authenticate button below, on the Security section and paste it there and click Authenticate, and now, you are ready to go!


    ## Deleting Objects
    We do our best to have all our URLs be
    [RESTful](http://en.wikipedia.org/wiki/Representational_state_transfer).
    Every endpoint (URL) may support one of four different http verbs. GET
    requests fetch information about an object, POST requests create objects,
    PUT requests update objects, and finally DELETE requests will delete
    objects.


    ## Structure

    ### The Envelope
    Every response is contained by an envelope. That is, each response has a
...
    ```json
    {
        "meta": {
...
        },
...

...

...
host: api.aswproject.fun
basePath: /
schemes:
...
produces:
...
consumes:
...
securityDefinitions:
  key:
    type: apiKey
    in: header
    name: access_token
security:
  - key: []
  
################################################################################
#                                   Parameters                                 #
################################################################################
...
paths:
  /users:
    get: 
      security: 
        - key: [] 
      tags: 
        - Users 
      description: |
        Get all users in the system 
      responses: 
        200: 
          description: OK 
          schema: 
            properties:
              meta:
                properties:
                  code:
                    type: integer

              data:
                title: users
                type: array
                items:
                  $ref: '#/definitions/User'

              _links:
                properties:
                  self:
                    type: object
                    properties:
                      href:
                        type: string
        401:
          $ref: '#/responses/Standard401ErrorResponse'
    post:
      tags:
        - Users
      description: |
        Create a new user
      parameters:
        - in: body
          name: User
          description: New user object
          required: false
          schema:
            $ref: "#/definitions/newUser"
      responses:
        201:
          description: Created
          schema:
            properties:
                meta:
                  properties:
                    code:
                      type: integer
                      
                data:
                  title: users
                  type: array
                  items:
                    $ref: '#/definitions/User'
                      
                _links:
                  properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
        403:
          $ref: '#/responses/Standard403ErrorResponse'
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
...
  newUser:
    type: object
    required:
      - name
      - email
      - password
      - password_confirmation
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
        format: password
      password_confirmation:
        type: string
        format: password
...
responses:
  Standard401ErrorResponse:
    description: Unauthorized
  Standard403ErrorResponse:
    description: Forbidden
...